LPD-72213 Extract from index
