+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/js/index.js
@@ -3,47 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {openConfirmModal} from 'frontend-js-components-web';
-import {getCheckedCheckboxes, postForm} from 'frontend-js-web';
-
-export function ExportImportManagementToolbarPropsTransformer({
-	portletNamespace,
-	...otherProps
-}) {
-	return {
-		...otherProps,
-		onActionButtonClick: (event, {item}) => {
-			if (item?.data?.action === 'deleteEntries') {
-				openConfirmModal({
-					message: Liferay.Language.get(
-						'are-you-sure-you-want-to-delete-the-selected-entries'
-					),
-					onConfirm: (isConfirmed) => {
-						if (isConfirmed) {
-							const form = document.getElementById(
-								`${portletNamespace}fm`
-							);
-
-							if (form) {
-								postForm(form, {
-									data: {
-										cmd: 'delete',
-										deleteBackgroundTaskIds:
-											getCheckedCheckboxes(
-												form,
-												`${portletNamespace}allRowIds`
-											),
-									},
-								});
-							}
-						}
-					},
-				});
-			}
-		},
-	};
-}
-
+export {ExportImportManagementToolbarPropsTransformer} from './components/ExportImportManagementToolbarPropsTransformer';
 export {NewExport} from '../revamp/js/pages/export/NewExport';
 export {ViewImportReportEntryDetail} from '../revamp/js/pages/import/report/ViewImportReportEntryDetail';
 export {default as formatDate} from '../revamp/js/utils/formatDate';
